+++ OpenSim/Common/ComponentConnector.h
@@ -262,7 +262,7 @@ private:
 };  // END class AbstractInput
 
 
-/** An Input<Y> must be connected by an Output<Y> */
+/** An Input<T> must be connected by an Output<T> */
 template<class T>
 class  Input : public AbstractInput {
     OpenSim_DECLARE_CONCRETE_OBJECT(Input, AbstractInput);
@@ -300,8 +300,53 @@ public:
 
 private:
     mutable SimTK::ReferencePtr< const Output<T>  > connectee;
-}; // END class Input<Y>
+}; // END class Input<T>
 
+
+/** A SimTK::Measure_ whose value is the value of an OpenSim::Input, and whose
+* dependsOn SimTK::Stage is the connectAt stage of the OpenSim::Input. This
+* Measure is useful for building OpenSim Components that use
+* a SimTK::Measure_ internally (e.g., Delay_).
+*/
+template <class T>
+class InputMeasure : public SimTK::Measure_<T> {
+public:
+    SimTK_MEASURE_HANDLE_PREAMBLE(InputMeasure, SimTK::Measure_<T>);
+
+    InputMeasure(SimTK::Subsystem& sub, const OpenSim::Input<T>& input)
+        :   SimTK::Measure_<T>(sub, new Implementation(input),
+            SimTK::AbstractMeasure::SetHandle()) {}
+
+    SimTK_MEASURE_HANDLE_POSTSCRIPT(InputMeasure, SimTK::Measure_<T>);
+};
+
+template <class T>
+class InputMeasure<T>::Implementation :
+        public SimTK::Measure_<T>::Implementation {
+public:
+    Implementation(const OpenSim::Input<T>& input)
+        :   SimTK::Measure_<T>::Implementation(), m_input(input) {}
+
+    Implementation* cloneVirtual() const override
+    {return new Implementation(*this);}
+    int getNumTimeDerivativesVirtual() const override {return 0;}
+    SimTK::Stage getDependsOnStageVirtual(int order) const override
+    { return m_input.getConnectAtStage(); }
+
+    // The meat of this class: return the Input's value.
+    void calcCachedValueVirtual(const SimTK::State& s,
+            int derivOrder, T& value) const override
+    {
+        SimTK_ASSERT1_ALWAYS(derivOrder == 0,
+                "InputMeasure::Implementation::calcCachedValueVirtual(): "
+                        "derivOrder %d seen but only 0 allowed.", derivOrder);
+
+        value = m_input.getValue(s);
+    }
+
+private:
+    const OpenSim::Input<T>& m_input;
+}; // END class InputMeasure<T>.
 
 } // end of namespace OpenSim