+++ OpenSim/Common/osimCommon.h
@@ -43,6 +43,7 @@
 #include "MultiplierFunction.h"
 #include "PolynomialFunction.h"
 
+#include "Delay.h"
 #include "ObjectGroup.h"
 #include "StorageInterface.h"
 #include "LoadOpenSimLibrary.h"